requirements.txt
openai==1.12.0
MetaTrader5
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
ollama
pytest

